add `BaseContext.decorate` interface
When extending contexts, it is common that a later context will want to
add functionality to an object defined via a previous context via a
decoration pattern. For example, a `LoggedInContext` might define a
`user`, and a `UserEditContext` might want to present that user for the
form via a Presenter object that uses a decoration pattern. In this
case, the `UserEditContext` could declare something like:

  decorate :user, decorator: 'UserForFormPresenter'

This will then do the equivalent of:

  def user
    UserForFormPresenter.new(@parent_context.user)
  end

This decorate interface can also be supplied with additional arguments,
like so:

  decorate :user, decorator: 'UserForFormPresenter', args: [:foo]

This requires that there be a `foo` method defined in the Context where
`decorate` is declared, and will do the equivalent of:

  def user
    UserForFormPresenter.new(@parent_context.user, foo)
  end