disallow overriding public methods in contexts
Prior to this commit, it was possible for contexts to override public
methods defined in previous contexts. This is an anti-pattern, because
one of the main goals of contexts is to tell a clear, linear story of
what is happening in a request. Overriding methods adds confusion, and
is a sign that contexts are being improperly used. This commit
explicitly prevents any context from overriding methods defined earlier
in the context chain. This happens at runtime when the `wrap` method is
called. If a method is found to be overwriting a previously-defined
public method, a `MethodOverrideError` exception is raised.

Two notes:

1. Methods that have been explicitly decorated using the `decorate`
   declaration are excluded from this constraint. Decoration is a valid
   use of contexts, and is not considered an anti-pattern. One of the
   reasons the `decorate` declaration was added before is that it
   signifies intent, and works within a clear and consistent interface
   pattern, which allows us to constrain the exception to that use case.

2. Non-public methods defined in a context are not subject to any of the
   constraints described above. A context can only access public methods
   of other contexts earlier in the chain, so those are the only ones
   that we care about constraining.