add method_missing logic to Context::Controller
This commit adds functionality to easily access context methods in
controllers. This is done through the use of method_missing.

Prior to this commit, if a controller wanted to make use of methods
defined in contexts, it would need to do so explicitly, like so:
  @__context.mymethod
Now, if `mymethod` is a public method available in the context stack,
the controller can simply call it directly without needing to
reference `@__context`.