rename @context to @__context in controller
It is possible that an application using this gem will want to use an
instance variable named `@context` in a controller for unrelated
purposes. If that were to happen, it would break this gem’s code.
This commit changes the variable name used in the
`Context::Controller` module to `@__context` to make that sort of
issue very unlikely.