fix gem dependencies
This commit fixes some gem dependencies that were copy/pasted from a
different gem when creating this repo initially.

The Rails dependency is set to > 4.0.2 because that’s when
`before_filter` was changed to `before_action`, and it’s unlikely that
anyone will need to use this gem for applications using versions of
Rails earlier than that.

There aren’t any tests that rely on ActiveRecord, so the AR
environement variable is not relevant.