fix naming bug in Context::BaseContextHelper
The code for this helper module was originally ported over from the
WegoWise app, where `context` was accessible via an attr_reader in
ApplicationController. But in this gem code, we (a) don’t have the
attr_reader and need to access the instance variable, and (b) changed
the instance variable name to `@__context` in commit 562d5a5. The
copy/pasted code wasn’t properly modified to account for this, making
this helper module useless.

This commit fixes the issue.